Refactor ObjectMapperUtils
