+++ src/main/java/com/alkemy/ong/services/impl/MemberServiceImpl.java
@@ -7,7 +7,7 @@ import com.alkemy.ong.models.MemberEntity;
 import com.alkemy.ong.repositories.IMemberRepository;
 import com.alkemy.ong.services.AWSS3Service;
 import com.alkemy.ong.services.MemberService;
-import com.alkemy.ong.services.mappers.ObjectMapperUtils;
+import com.alkemy.ong.utils.ObjectMapperUtils;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageImpl;
 import org.springframework.data.domain.Pageable;
