+++ src/main/java/com/alkemy/ong/services/impl/TestimonialsServiceImpl.java
@@ -6,7 +6,7 @@ import com.alkemy.ong.models.TestimonialsEntity;
 import com.alkemy.ong.repositories.ITestimonialsRepository;
 import com.alkemy.ong.services.AWSS3Service;
 import com.alkemy.ong.services.TestimonialsService;
-import com.alkemy.ong.services.mappers.ObjectMapperUtils;
+import com.alkemy.ong.utils.ObjectMapperUtils;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.data.domain.Page;
